requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
distro==1.9.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
numpy==2.4.3
openai==2.26.0
portalocker==3.2.0
protobuf==7.34.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
qdrant-client==1.17.0
rank-bm25==0.2.2
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
